+++ npm/create-cypress-tests/__snapshots__/init-component-testing.test.ts.js
@@ -1,6 +1,5 @@
 exports['Injected overridden webpack template cypress.json'] = `
 {
-  "experimentalComponentTesting": true,
   "componentFolder": "cypress/component",
   "testFiles": "**/*.spec.{js,ts,jsx,tsx}"
 }
@@ -10,8 +9,11 @@ exports['Injected overridden webpack template plugins/index.js'] = `
 const injectDevServer = require("@cypress/react/plugins/react-scripts");
 
 module.exports = (on, config) => {
-  injectDevServer(on, config);
-  return config; // IMPORTANT to return the config object
+  if (config.testingType === "component") {
+    injectDevServer(on, config);
+  }
+
+  return config; // IMPORTANT to return a config
 };
 
 `
@@ -22,7 +24,6 @@ import "./commands.js";
 
 exports['injects guessed next.js template cypress.json'] = `
 {
-  "experimentalComponentTesting": true,
   "componentFolder": "src",
   "testFiles": "**/*.spec.{js,ts,jsx,tsx}"
 }
@@ -32,8 +33,11 @@ exports['injects guessed next.js template plugins/index.js'] = `
 const injectDevServer = require("@cypress/react/plugins/next");
 
 module.exports = (on, config) => {
-  injectDevServer(on, config);
-  return config; // IMPORTANT to return the config object
+  if (config.testingType === "component") {
+    injectDevServer(on, config);
+  }
+
+  return config; // IMPORTANT to return a config
 };
 
 `