+++ npm/eslint-plugin-dev/lib/index.js
@@ -50,6 +50,8 @@ const baseRules = {
     'error',
     2,
     {
+      // TODO: fix this, we shouldn't need to ignore TemplateLiterals
+      'ignoredNodes': ['TemplateLiteral'],
       'SwitchCase': 1,
       'MemberExpression': 0,
     },
@@ -193,7 +195,8 @@ const baseRules = {
   ],
   'space-infix-ops': 'error',
   'space-unary-ops': 'error',
-  'template-curly-spacing': 'error',
+  // TODO: change this back to 'error'
+  'template-curly-spacing': 'off',
   'use-isnan': 'error',
   'valid-typeof': 'error',
 }
@@ -255,7 +258,10 @@ module.exports = {
           },
         },
         {
-          files: '*.ts*',
+          files: [
+            '*.ts',
+            '*.tsx',
+          ],
           parser: '@typescript-eslint/parser',
           plugins: [
             '@typescript-eslint',