+++ npm/vite-dev-server/cypress/plugins.js
@@ -1,7 +1,15 @@
-import { startDevServer } from '@cypress/vite-dev-server'
+const path = require('path')
+const { startDevServer } = require('../dist')
 
 module.exports = (on, config) => {
-  on('dev-server:start', async (options) => startDevServer({ options }))
+  on('dev-server:start', async (options) => {
+    return startDevServer({
+      options,
+      viteConfig: {
+        configFile: path.resolve(__dirname, '..', 'vite.config.ts'),
+      },
+    })
+  })
 
   return config
 }