+++ npm/webpack-dev-server/src/aut-runner.ts
@@ -1,18 +1,7 @@
-/* eslint-disable no-console */
-/*eslint-env browser */
+/* eslint-env browser */
 
-function appendTargetIfNotExists (id: string, tag = 'div', parent = document.body) {
-  let node = document.getElementById(id)
-
-  node = document.createElement(tag)
-  node.setAttribute('id', id)
-  parent.appendChild(node)
-
-  return node
-}
-
-export function init (importPromises, parent = (window.opener || window.parent)) {
-  const Cypress = (window as any).Cypress = parent.Cypress
+export function init (importPromises, parent: Window = (window.opener || window.parent)) {
+  const Cypress = window.Cypress = parent.Cypress
 
   if (!Cypress) {
     throw new Error('Tests cannot run without a reference to Cypress!')
@@ -20,27 +9,4 @@ export function init (importPromises, parent = (window.opener || window.parent))
 
   Cypress.onSpecWindow(window, importPromises)
   Cypress.action('app:window:before:load', window)
-
-  // In this variable, we save head
-  // innerHTML to account for loader installed styles
-  let headInnerHTML = ''
-
-  // before the run starts save
-  Cypress.on('run:start', () => {
-    headInnerHTML = document.head.innerHTML
-  })
-
-  // Before all tests we are mounting the root element, **not beforeEach**
-  // Cleaning up platform between tests is the responsibility of the specific adapter
-  // because unmounting react/vue component should be done using specific framework API
-  // (for devtools and to get rid of global event listeners from previous tests.)
-  Cypress.on('test:before:run', () => {
-    document.body.innerHTML = ''
-    document.head.innerHTML = headInnerHTML
-    appendTargetIfNotExists('__cy_root')
-  })
-
-  return {
-    restartRunner: Cypress.restartRunner,
-  }
 }