+++ packages/desktop-gui/cypress/support/index.js
@@ -4,7 +4,6 @@ require('@packages/ui-components/cypress/support/customPercyCommand')({
   },
 })
 
-require('cypress-react-unit-test/dist/hooks')
 require('cypress-real-events/support')
 
 const BluebirdPromise = require('bluebird')