+++ packages/runner-ct/src/lib/useHotKey.ts
@@ -1,10 +1,11 @@
 import * as React from 'react'
 import hotkeys from 'hotkeys-js'
 
-export function useGlobalHotKey (shortcut: string, handler: () => void) {
+export const useGlobalHotKey = (shortcut: string, handler: () => void) => {
   React.useEffect(() => {
-    hotkeys(shortcut, handler)
+    // Provide a clean handler since we assert that `handler` has no args
+    hotkeys(shortcut, () => handler())
 
     return () => hotkeys.unbind(shortcut)
-  }, [shortcut])
+  }, [shortcut, handler])
 }