+++ packages/runner/cypress/fixtures/errors/uncaught_spec.js
@@ -1,6 +1,18 @@
+import Bluebird from 'bluebird'
+
 import './setup'
 
 describe('uncaught errors', { defaultCommandTimeout: 0 }, () => {
+  it('sync app visit exception', () => {
+    cy.visit('/index.html')
+    cy.get('.trigger-sync-error').click()
+  })
+
+  it('sync app navigates to visit exception', () => {
+    cy.visit('/index.html')
+    cy.get('.go-to-visit-error').click()
+  })
+
   it('sync app exception', () => {
     cy.visit('/index.html')
     cy.get('.trigger-sync-error').click()
@@ -12,7 +24,13 @@ describe('uncaught errors', { defaultCommandTimeout: 0 }, () => {
     cy.wait(10000)
   })
 
-  it('async exception', () => {
+  it('app unhandled rejection', () => {
+    cy.visit('/index.html')
+    cy.get('.trigger-unhandled-rejection').click()
+    cy.wait(10000)
+  })
+
+  it('async spec exception', () => {
     setTimeout(() => {
       ({}).bar()
     })
@@ -20,9 +38,44 @@ describe('uncaught errors', { defaultCommandTimeout: 0 }, () => {
     cy.wait(10000)
   })
 
-  it('async exception with done', (done) => {
+  // eslint-disable-next-line mocha/handle-done-callback
+  it('async spec exception with done', (done) => {
     setTimeout(() => {
       ({}).bar()
     })
+  })
+
+  it('spec unhandled rejection', () => {
+    Promise.reject(new Error('Unhandled promise rejection from the spec'))
+
+    cy.wait(10000)
+  })
+
+  // eslint-disable-next-line mocha/handle-done-callback
+  it('spec unhandled rejection with done', (done) => {
+    Promise.reject(new Error('Unhandled promise rejection from the spec'))
+  })
+
+  it('spec Bluebird unhandled rejection', () => {
+    Bluebird.reject(new Error('Unhandled promise rejection from the spec'))
+
+    cy.wait(10000)
+  })
+
+  // eslint-disable-next-line mocha/handle-done-callback
+  it('spec Bluebird unhandled rejection with done', (done) => {
+    Bluebird.reject(new Error('Unhandled promise rejection from the spec'))
+  })
+
+  // TODO: Cypress.Promise.reject() gets caught by AUT. Can/should
+  // we handle that somehow?
+
+  it('exception inside uncaught:exception', () => {
+    cy.on('uncaught:exception', () => {
+      ({}).bar()
+    })
+
+    cy.visit('/index.html')
+    cy.get('.trigger-sync-error').click()
   })
 })